+++ sdk/go/examples/secrets/main.go
@@ -1,0 +1,15 @@
+/*
+ * This example is part of the Modus project, licensed under the Apache License 2.0.
+ * You may modify and use this example in accordance with the license.
+ * See the LICENSE file that accompanied this code for further details.
+ */
+
+package main
+
+import (
+	"github.com/hypermodeinc/modus/sdk/go/pkg/secrets"
+)
+
+func GetSecretValue(name string) (string, error) {
+	return secrets.GetSecretValue(name)
+}