+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # Change Log
 
+## UNRELEASED
+
+- feat: add kubernetes secrets provider and API to read secrets [#885](https://github.com/hypermodeinc/modus/pull/885)
+
 ## 2025-06-10 - Runtime 0.18.0-alpha.6
 
 - fix: address agent lifecycle issues [#881](https://github.com/hypermodeinc/modus/pull/881)